pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "resume-tailor"
version = "0.1.0"
description = "A deterministic data pipeline for tailoring resumes to job descriptions"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["resume", "llm", "latex", "automation", "pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Text Processing :: Markup :: LaTeX",
]
dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "jinja2>=3.1.2",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
]

[project.optional-dependencies]
ollama = ["langchain-ollama>=0.0.1"]
groq = ["langchain-groq>=0.0.1"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
resume-tailor = "src.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/resume-tailor"
Repository = "https://github.com/yourusername/resume-tailor"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
